Fix case-sensitive filename issue and add AgreementController.php
