+++ app/Http/Controllers/Agreement/AgreementController.php
@@ -8,9 +8,14 @@ use App\Models\GrieMoaAcademicPelaporanPartner;
 use App\Models\GrieMoaAcademicPelaporanProdi;
 use App\Models\GrieMoaAcademicPelaporanScope;
 use App\Models\GrieMoaAcademic;
+use App\Models\GrieMoaAcademicFaculty;
+use App\Models\GrieMoaAcademicPartner;
+use App\Models\GrieMoaAcademicProdi;
+use App\Models\GrieMoaAcademicScope;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 
 class AgreementController extends Controller
 {
@@ -18,35 +23,85 @@ class AgreementController extends Controller
 
         $univ = DB::table('m_university')
         ->get();
-
         $unit = DB::table('m_fakultas_unit')
         ->get(); 
-        
         $country = DB::table('m_country')
         ->get();
-
         $scope = DB::table('m_collaboration_scope')
         ->get();
-
         $department = DB::table('m_departemen')
         ->get();
+        $prodi = DB::table('m_prodi')
+        ->get();
+        $data = $id ? GrieMoaAcademicPelaporan::findOrFail($id) : null;
 
+        if ($data !== null) {
+                $selPartners = GrieMoaAcademicPelaporanPartner::where('id_moa_academic', $data->id)->get()->toArray();
+                $selPartners = array_column($selPartners, 'id_partner_university'); 
+                $selFaculties = GrieMoaAcademicPelaporanFaculty::where('id_moa_academic', $data->id)->get()->toArray();
+                $selFaculties = array_column($selFaculties, 'id_faculty'); 
+                $selScopes = GrieMoaAcademicPelaporanScope::where('id_moa_academic', $data->id)->get()->toArray();
+                $selScopes = array_column($selScopes, 'id_collaboration_scope'); 
+                $selProdis = GrieMoaAcademicPelaporanProdi::where('id_moa_academic', $data->id)->get()->toArray();
+                $selProdis = array_column($selProdis, column_key: 'id_program_study_unair'); 
+                return view('agreement.form_pelaporan', compact('data',  'selPartners', 'selFaculties','selScopes', 'selProdis', 'univ', 'unit','country', 'scope', 'department', 'prodi'));
+            }
+        else{ 
+            return view('agreement.form_pelaporan', compact('data',  'univ', 'unit','country', 'scope', 'department', 'prodi'));}
+    }
+
+    public function tambah_master_database($id = null){
+
+        $univ = DB::table('m_university')
+        ->get();
+        $unit = DB::table('m_fakultas_unit')
+        ->get(); 
+        $country = DB::table('m_country')
+        ->get();
+        $scope = DB::table('m_collaboration_scope')
+        ->get();
+        $department = DB::table('m_departemen')
+        ->get();
         $prodi = DB::table('m_prodi')
         ->get();
 
-        $data = $id ? GrieMoaAcademicPelaporan::findOrFail($id) : null;
+        $data = $id ? GrieMoaAcademic::findOrFail($id) : null;
+        if ($data !== null) {
+            $selPartners = GrieMoaAcademicPartner::where('id_moa_academic', $data->id)->get()->toArray();
+            $selPartners = array_column($selPartners, 'id_partner_university'); 
+            $selFaculties = GrieMoaAcademicFaculty::where('id_moa_academic', $data->id)->get()->toArray();
+            $selFaculties = array_column($selFaculties, 'id_faculty'); 
+            $selScopes = GrieMoaAcademicScope::where('id_moa_academic', $data->id)->get()->toArray();
+            $selScopes = array_column($selScopes, 'id_collaboration_scope'); 
+            $selProdis = GrieMoaAcademicProdi::where('id_moa_academic', $data->id)->get()->toArray();
+            $selProdis = array_column($selProdis, column_key: 'id_program_study_unair'); 
 
-        return view('agreement.form_pelaporan', compact('data', 'univ', 'unit','country', 'scope', 'department', 'prodi'));
+            return view('agreement.form_master_database', compact('data', 'selPartners', 'selFaculties', 'selScopes', 'selProdis', 'univ', 'unit','country', 'scope', 'department', 'prodi'));
+        }else{
+                return view('agreement.form_master_database', compact('data', 'univ', 'unit','country', 'scope', 'department', 'prodi'));
     }
+
+        }
+
+    
 
 
 public function destroy_pelaporan($id){
-        $del = GrieMoaAcademicPelaporan::where("ID", $id)->first();
-
-        if ($del) {
+    
+        try {
+            $del = GrieMoaAcademicPelaporan::where("id", $id)->first();
             $del->delete();  
-            return redirect()->route('view_pelaporan');
-        }        
+    
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Data berhasil dihapus!'
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Gagal menghapus data.'
+            ]);
+        }
 } 
 
 public function store_pelaporan(Request $request, $id = null) {
@@ -61,13 +116,21 @@ public function store_pelaporan(Request $request, $id = null) {
 
     if ($id) {
         $pelaporan = GrieMoaAcademicPelaporan::findOrFail($id); 
+
+        GrieMoaAcademicPelaporanFaculty::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicPelaporanProdi::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicPelaporanScope::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicPelaporanPartner::where("id_moa_academic", $id)->delete();
+
     } else {
         $pelaporan = new GrieMoaAcademicPelaporan();
+        $pelaporan->approval_pelaporan = 0;
+        $pelaporan->created_by = Auth::user()->id;
     }
 
-    // Assign nilai ke model
-    $pelaporan->id_dosen = 0;
-    $pelaporan->id_partner_university = 0;
+    $pelaporan->id_dosen = null;
+    $pelaporan->created_date = now();
+    $pelaporan->id_partner_university = null;
     $pelaporan->tipe_moa = $request->input('jenisP');
     $pelaporan->kategori_tridharma = $request->input('triDharma');
     $pelaporan->kategori = $request->input('jenisP');
@@ -77,7 +140,19 @@ public function store_pelaporan(Request $request, $id = null) {
     $pelaporan->text_end_date = "Automatically Renewed";
     $pelaporan->jenis_naskah = $request->input('docP');
     $pelaporan->title = $request->input('tittleP');
-    $pelaporan->link_download_naskah = $request->input('linkDownload');
+    // $pelaporan->link_download_naskah = $request->input('linkDownload');
+    if ($request->hasFile('linkDownload')) {
+        $file = $request->file('linkDownload');
+        $storagePath = '/naskah';
+    
+        if (!Storage::disk('outside')->exists($storagePath)) {
+            Storage::disk('outside')->makeDirectory($storagePath);
+        }
+        $fileName = uniqid() . '_' . $file->getClientOriginalName();
+        $file->storeAs($storagePath, $fileName, 'outside');
+        $pelaporan->link_download_naskah = $storagePath . '/' . $fileName;
+    }
+
     $pelaporan->mou_start_date = $request->input('startDate');
     $pelaporan->mou_end_date = $request->input('endDate');
     $pelaporan->id_department_unair = $request->input('deptP');
@@ -88,6 +163,7 @@ public function store_pelaporan(Request $request, $id = null) {
     $pelaporan->type_institution_partner = $request->input('typeP');
     $pelaporan->signatories_unair_name = $request->input('nosUnair');
     $pelaporan->signatories_unair_pos = $request->input('nopUnair');
+    $pelaporan->current_role = 'KSLN';
     $pelaporan->signatories_partner_name = $request->input('nosPart');
     $pelaporan->signatories_partner_pos = $request->input('nopPart');
     $pelaporan->pic_mitra_nama = $request->input('namePic');
@@ -98,7 +174,7 @@ public function store_pelaporan(Request $request, $id = null) {
     $pelaporan->pic_fak_jabatan = $request->input('postPic2');
     $pelaporan->pic_fak_email = $request->input('emailPic2');
     $pelaporan->pic_fak_phone = $request->input('telpPic2');
-    $pelaporan->approval_pelaporan = 0;
+
 
     if ($request->input('jenisP') == "Riset") {
         $pelaporan->source_funding = $request->input('sourceFund');
@@ -138,8 +214,6 @@ public function store_pelaporan(Request $request, $id = null) {
             ['id_moa_academic' => $pelaporan_id, 'id_program_study_unair' => (int) $programStudyId]
         );
     }
-
-    session()->flash('success', 'Data berhasil disimpan!');
     return redirect()->route('view_pelaporan');
 }
 
@@ -155,7 +229,7 @@ public function store_pelaporan(Request $request, $id = null) {
             DB::table('grie_moa_academic_pelaporan_partner')
                 ->select(
                     'id_moa_academic',
-                    DB::raw('GROUP_CONCAT(u.name) AS partner')
+                    DB::raw('STRING_AGG(u.name, \', \') AS partner') 
                 )
                 ->leftjoin('m_university as u', 'u.id', '=', 'grie_moa_academic_pelaporan_partner.id_partner_university')
                 ->groupBy('id_moa_academic')
@@ -166,6 +240,9 @@ public function store_pelaporan(Request $request, $id = null) {
             't.id'
         )
         ->leftjoin('m_fakultas_unit as f', 'f.id', '=', 't.id_fakultas')
+        ->when(!Auth::user()->hasRole('gpc'), function ($query) {
+            $query->where('t.created_by', '=', Auth::user()->id);
+        })
         ->orderBy('created_date', 'desc')
         ->get();
 
@@ -173,10 +250,59 @@ public function store_pelaporan(Request $request, $id = null) {
     }
 
     public function approve_pelaporan(Request $request, $id){
+        
         $pelaporan = GrieMoaAcademicPelaporan::find($id);
+    
+        $partnerId = GrieMoaAcademicPelaporanPartner::where('id_moa_academic', $pelaporan->id)->get()->toArray();
+        $facultyId = GrieMoaAcademicPelaporanFaculty::where('id_moa_academic', $pelaporan->id)->get()->toArray();
+        $programStudyId = GrieMoaAcademicPelaporanProdi::where('id_moa_academic', $pelaporan->id)->get()->toArray();
+        $scopeId = GrieMoaAcademicPelaporanScope::where('id_moa_academic', $pelaporan->id)->get()->toArray();
+
         $pelaporan->approval_pelaporan = 1;
         $pelaporan->approval_status = null;
         $pelaporan->save();
+        
+        $newMoa = new GrieMoaAcademic;
+        
+        $newMoa->fill($pelaporan->getAttributes());
+        $newMoa->status = 'Completed';
+        $newMoa->status_lapkerma = 'BELUM';
+        $newMoa->age_archive_sn = GrieMoaAcademic::generateNumber();
+        $newMoa->lapkerma_archive = $newMoa->age_archive_sn.'.KSLN';
+        $newMoa->year = date('Y', strtotime($newMoa->mou_start_date));
+        $new_moa_id = $newMoa->id;
+
+        $newMoa->save();
+            $new_moa_id = $newMoa->id;
+            foreach ($partnerId as $partnerIds) {
+                GrieMoaAcademicPartner::updateOrCreate(
+                    ['id_moa_academic' => $new_moa_id, 'id_partner_university' => (int) $partnerIds['id_partner_university']],
+                    ['id_moa_academic' => $new_moa_id, 'id_partner_university' => (int) $partnerIds['id_partner_university']]
+                );
+            }
+            
+            foreach ($scopeId as $scopeIds) {
+                GrieMoaAcademicScope::updateOrCreate(
+                    ['id_moa_academic' => $new_moa_id, 'id_collaboration_scope' => (int) $scopeIds['id_collaboration_scope']],
+                    ['id_moa_academic' => $new_moa_id, 'id_collaboration_scope' => (int) $scopeIds['id_collaboration_scope']]
+                );
+            }
+            
+            foreach ($facultyId as $facultyIds) {
+                GrieMoaAcademicFaculty::updateOrCreate(
+                    ['id_moa_academic' => $new_moa_id, 'id_faculty' => (int) $facultyIds['id_faculty']],
+                    ['id_moa_academic' => $new_moa_id, 'id_faculty' => (int) $facultyIds['id_faculty']]
+                );
+            }
+            
+        
+            foreach ($programStudyId as $programStudyIds) {
+                GrieMoaAcademicProdi::updateOrCreate(
+                    ['id_moa_academic' => $new_moa_id, 'id_program_study_unair' => (int) $programStudyIds['id_program_study_unair']],
+                    ['id_moa_academic' => $new_moa_id, 'id_program_study_unair' => (int) $programStudyIds['id_program_study_unair']]
+                );
+            }
+
         return redirect()->route('view_pelaporan');
     }
 
@@ -207,18 +333,19 @@ public function store_pelaporan(Request $request, $id = null) {
             ->where(function($query) {
                 $query->where('g.status', '=', 'Completed')
                       ->orWhere('g.status', '=', 'Progress');
-            })  // Kondisi untuk status Completed atau Progress
-            ->whereNotNull('g.link_download_naskah')  // Kondisi untuk memastikan link_download_naskah tidak null
+            })  
+            ->whereNotNull('g.link_download_naskah')  
             ->where('g.link_download_naskah', '!=', '')  // K
+            ->orderByDesc('lapkerma_archive')
             ->get();
 
         $result = DB::table('grie_moa_academic_scope as gs')
                 ->select(
                     'gs.id_moa_academic',
-                    DB::raw('GROUP_CONCAT(DISTINCT u2.name) AS partner_involved'),
-                    DB::raw('GROUP_CONCAT(DISTINCT p.name_eng) AS prodi_involved'),
-                    DB::raw('GROUP_CONCAT(DISTINCT fu.nama_eng) AS faculty_involved'),
-                    DB::raw('GROUP_CONCAT(DISTINCT cs.name) AS collaboration_scope')
+                    DB::raw('STRING_AGG(DISTINCT u2.name, \', \') AS partner_involved'),
+                    DB::raw('STRING_AGG(DISTINCT p.name_eng, \', \') AS prodi_involved'),
+                    DB::raw('STRING_AGG(DISTINCT fu.nama_eng, \', \') AS faculty_involved'),
+                    DB::raw('STRING_AGG(DISTINCT cs.name, \', \') AS collaboration_scope')
                 )
                 ->leftJoin('grie_moa_academic_prodi as gap', 'gap.id_moa_academic', '=', 'gs.id_moa_academic')
                 ->leftJoin('m_prodi as p', 'p.id', '=', 'gap.id_program_study_unair')
@@ -242,25 +369,209 @@ public function store_pelaporan(Request $request, $id = null) {
                 
                     return $item;
                 });
-            
+
         
         return view('agreement.view_database', compact( 'merged'));
     }
 
     public function destroy_database_agreement($id){
-        $del = GrieMoaAcademic::where("ID", $id)->first();
+       
 
-        if ($del) {
+        try {
+            $del = GrieMoaAcademic::where("id", $id)->first();
             $del->delete();  
-            return redirect()->route('view_database');
-        }        
+    
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Data berhasil dihapus!'
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Gagal menghapus data.'
+            ]);
+        }
 
     }
 
+    public function upload_bukti($id){
+
+        $data = GrieMoaAcademic::find($id);
+        return view('agreement.upload_pelaporan', compact('data'));
+
+    }
+
+
+    public function store_bukti(Request $request, $id){
+
+        $agreement = GrieMoaAcademic::find($id);
+
+        if ($request->hasFile('linkPelaporan')) {
+            $file = $request->file('linkPelaporan');
+            $storagePath = '/naskah';
+    
+            // Pastikan direktori penyimpanan ada
+            if (!Storage::disk('outside')->exists($storagePath)) {
+                Storage::disk('outside')->makeDirectory($storagePath);
+            }
+    
+            // Simpan file dengan nama unik
+            $fileName = uniqid() . '_' . $file->getClientOriginalName();
+            $file->storeAs($storagePath, $fileName, 'outside');
+    
+            // Perbarui hanya kolom link_pelaporan
+            $agreement->update([
+                'link_pelaporan' => $storagePath . '/' . $fileName,
+            ]);
+        }
+       
+
+        return redirect()->route('view_database');
+
+    }
+
+    public function generate_number(){
+        $year = date('Y');
+        $lastRecord = DB::table('grie_moa_academic')
+            ->where('age_archive_sn', 'like', $year . '%')
+            ->orderBy('age_archive_sn', 'desc')
+            ->first();
+    
+        if ($lastRecord && isset($lastRecord->age_archive_sn)) {
+            $lastNumber = explode('.', $lastRecord->age_archive_sn)[1];
+            $newNumber = $lastNumber + 1;
+        } else {
+            $newNumber = 1;
+        }
+    
+        $result = $year . '.' . str_pad($newNumber, 3, '0', STR_PAD_LEFT);
+        return $result;
+    }
+
+    public function store_master_database(Request $request, $id = null) {
+
+    $country = DB::table("m_country")
+        ->where("id", "=", $request->input('countryP'))
+        ->pluck("m_country.name")->first();
+
+    $dept = DB::table("m_departemen")
+        ->where("id", "=", $request->input('deptP'))
+        ->pluck("nama_ind")->first();
+    
+
+    if ($id) {
+        $agreement = GrieMoaAcademic::findOrFail($id); 
+        GrieMoaAcademicFaculty::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicProdi::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicScope::where("id_moa_academic", $id)->delete();
+        GrieMoaAcademicPartner::where("id_moa_academic", $id)->delete();
+
+    } else {
+        $agreement = new GrieMoaAcademic();
+        $agreement->created_by = Auth::user()->id;
+        if(Auth::user()->hasRole('gpc')){
+            $agreement->current_role = 'KSLN';
+            $agreement->current_iterasi = 1;
+        }
+    }
+    $agreement->current_id_status = 0;
+    $agreement->id_dosen = null;
+    $agreement->created_date = now();
+    $agreement->id_partner_university = null;
+    $agreement->tipe_moa = $request->input('jenisP');
+    $agreement->kategori_tridharma = $request->input('triDharma');
+    $agreement->kategori = $request->input('jenisP');
+    $agreement->id_country = $request->input('countryP');
+    $agreement->text_country = $country;
+    $agreement->id_fakultas = $request->input('unitP');
+    $agreement->text_end_date = "Automatically Renewed";
+    $agreement->jenis_naskah = $request->input('docP');
+    $agreement->title = $request->input('tittleP');
+    if ($request->hasFile('linkDownload')) {
+        $file = $request->file('linkDownload');
+        $storagePath = '/naskah';
+    
+        if (!Storage::disk('outside')->exists($storagePath)) {
+            Storage::disk('outside')->makeDirectory($storagePath);
+        }
+        $fileName = uniqid() . '_' . $file->getClientOriginalName();
+        $file->storeAs($storagePath, $fileName, 'outside');
+        $agreement->link_download_naskah = $storagePath . '/' . $fileName;
+    }
+    $agreement->mou_start_date = $request->input('startDate');
+    $agreement->mou_end_date = $request->input('endDate');
+    $agreement->id_department_unair = $request->input('deptP');
+    $agreement->text_department_unair = $dept;
+    $agreement->department_partner = $request->input('partDept');
+    $agreement->faculty_partner = $request->input('partnerFac');
+    $agreement->program_study_partner = $request->input('partnerStuProg');
+    $agreement->type_institution_partner = $request->input('typeP');
+    $agreement->signatories_unair_name = $request->input('nosUnair');
+    $agreement->signatories_unair_pos = $request->input('nopUnair');
+    $agreement->signatories_partner_name = $request->input('nosPart');
+    $agreement->signatories_partner_pos = $request->input('nopPart');
+    $agreement->pic_mitra_nama = $request->input('namePic');
+    $agreement->pic_mitra_jabatan = $request->input('postPic');
+    $agreement->pic_mitra_email = $request->input('emailPic');
+    $agreement->pic_mitra_phone = $request->input('telpPic');
+    $agreement->pic_fak_nama = $request->input('namePic2');
+    $agreement->pic_fak_jabatan = $request->input('postPic2');
+    $agreement->pic_fak_email = $request->input('emailPic2');
+    $agreement->pic_fak_phone = $request->input('telpPic2');
+    $agreement->is_queue = $request->input('queueP');
+    $agreement->status_lapkerma = $request->input('statusLapkerma');
+    $agreement->category_document = $request->input(key: 'catNaskah');
+    $agreement->skema = $request->input('skema');
+    $agreement->status = 'Completed';
+    $agreement->status_lapkerma = 'BELUM';
+    $agreement->age_archive_sn = GrieMoaAcademic::generateNumber();        
+    $agreement->lapkerma_archive = $agreement->age_archive_sn.'.KSLN';
+    $agreement->year = date('Y', strtotime($agreement->mou_start_date));
+    // $agreement->link_pelaporan = 'None';
+    // $agreement->current_id_status = 0;
+
+
+    $agreement->save();
+    $agreement_id = $agreement->id;
+
+    foreach ($request->input('partnerP', []) as $partnerId) {
+        GrieMoaAcademicPartner::updateOrCreate(
+            ['id_moa_academic' => $agreement_id, 'id_partner_university' => (int) $partnerId],
+            ['id_moa_academic' => $agreement_id, 'id_partner_university' => (int) $partnerId]
+        );
+    }
+
+    foreach ($request->input('scopeP', []) as $scopeId) {
+        GrieMoaAcademicScope::updateOrCreate(
+            ['id_moa_academic' => $agreement_id, 'id_collaboration_scope' => (int) $scopeId],
+            ['id_moa_academic' => $agreement_id, 'id_collaboration_scope' => (int) $scopeId]
+        );
+    }
+
+    foreach ($request->input('FacP', []) as $facultyId) {
+        GrieMoaAcademicFaculty::updateOrCreate(
+            ['id_moa_academic' => $agreement_id, 'id_faculty' => (int) $facultyId],
+            ['id_moa_academic' => $agreement_id, 'id_faculty' => (int) $facultyId]
+        );
+    }
+
+    foreach ($request->input('stuProgP', []) as $programStudyId) {
+        GrieMoaAcademicProdi::updateOrCreate(
+            ['id_moa_academic' => $agreement_id, 'id_program_study_unair' => (int) $programStudyId],
+            ['id_moa_academic' => $agreement_id, 'id_program_study_unair' => (int) $programStudyId]
+        );
+    }
+
+    session()->flash('success', 'Data berhasil disimpan!');
+    return redirect()->route('view_database');
+
+
+    }
     public function email_list(){
         
         return view('agreement.email_list');
     }
+
 
     public function review_agreement(){
